Move provider functionality
