+++ src/scheduling/simple_schedule.py
@@ -1,8 +1,12 @@
 import numpy as np
 import pulp
 
+from src.common import CircuitJob, ScheduledJob
+from src.provider import Accelerator
+
 from .calculate_makespan import calculate_makespan
-from .solve_lp import solve_lp
+from .setup_lp import _set_up_base_lp
+from .solve_lp import solve_lp, _solve_lp
 from .types import JobResultInfo, LPInstance, PTimes, STimes
 
 
@@ -72,3 +76,70 @@ def _get_simple_setup_times(
     for times in new_times:
         del times[0]
     return new_times
+
+
+def generate_simple_schedule_provider(
+    jobs: list[CircuitJob],
+    accelerators: list[Accelerator],
+    big_m: int = 1000,
+    t_max: int = 2**7,
+    **kwargs,
+) -> list[ScheduledJob]:
+    """Calclulate a schedule for the given jobs and accelerators based on the simple MILP.
+
+    The simple MILP includes the machine dependent setup times.
+    Args:
+        jobs (list[CircuitJob]): The list of jobs to run.
+        accelerators (list[Accelerator]): The list of available accelerators.
+        big_m (int, optional): M hepler for LP. Defaults to 1000.
+        t_max (int, optional): Max number of Timesteps. Defaults to 2**7.
+    Returns:
+        list[ScheduledJob]: A list of Jobs scheduled to accelerators.
+    """
+    lp_instance = _set_up_base_lp(jobs, accelerators, big_m, list(range(t_max)))
+    # (4) - (7), (9)
+    p_times = pulp.makeDict(
+        [lp_instance.jobs[1:], lp_instance.machines],
+        _get_processing_times(jobs, accelerators),
+        0,
+    )
+    s_times = pulp.makeDict(
+        [lp_instance.jobs[1:], lp_instance.machines],
+        _get_simple_setup_times_provider(jobs, accelerators),
+        0,
+    )
+
+    for job in lp_instance.jobs[1:]:
+        lp_instance.problem += lp_instance.c_j[job] >= lp_instance.s_j[  # (7)
+            job
+        ] + pulp.lpSum(
+            lp_instance.x_ik[job][machine]
+            * (p_times[job][machine] + s_times[job][machine])
+            for machine in lp_instance.machines
+        )
+
+    return _solve_lp(lp_instance, jobs, accelerators)
+
+
+def _get_simple_setup_times_provider(
+    base_jobs: list[CircuitJob], accelerators: list[Accelerator]
+) -> list[list[float]]:
+    return [
+        [
+            qpu.compute_setup_time(job_i.instance, circuit_to=None)
+            for qpu in accelerators
+        ]
+        for job_i in base_jobs
+        if job_i.instance is not None
+    ]
+
+
+def _get_processing_times(
+    base_jobs: list[CircuitJob],
+    accelerators: list[Accelerator],
+) -> list[list[float]]:
+    return [
+        [qpu.compute_processing_time(job.instance) for qpu in accelerators]
+        for job in base_jobs
+        if job.instance is not None
+    ]
